Rails 5 throws an exception if you set a non-existent callback to skip
Which previous versions didn't mind. With Factory Girl and the rspec tests
when building Dataset and DatasetFiles, as we don't want github being
called, the callbacks are skipped.

As they are skipped once, the next time the object is built, Rails 5
throws an exception saying that the call back doesn't exist.

Now, we only skip callbacks when creating objects if the callback
actually exists!