Rails 5 update with the org_dataset_ids column produces a mismatch of types
In my environment (Postgres 9.5) when the org_dataset_ids are saved as
an array of integers (in an text/array column) they are returned as an
array of string integers. This causes a test to fail.

On further investigation, it looks like production Octopub is running on
Postgres 9.3 which may explain the discrepancy.

Also, it looks like the column, whilst populated, was added in June 2016
and isn't actually used by the application.

It's possible therefore that it should actually be removed!

For now, the test converts the array of strings to an array of ints and
now passes